# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ShareUploadRequestContentType = typeof ShareUploadRequestContentType[keyof typeof ShareUploadRequestContentType];


export const ShareUploadRequestContentType = {
  'audio/mpeg': 'audio/mpeg',
  'audio/mp3': 'audio/mp3',
  'audio/wav': 'audio/wav',
} as const;

export interface ShareUploadRequest {
  name: string;
  /**
     * @minimum 1
     * @maximum 52428800
     */
  size: number;
  contentType: ShareUploadRequestContentType;
}

export interface ShareUploadResponse {
  uploadURL: string;
  objectPath: string;
  expiresInSeconds: number;
}

export interface SharedAudio {
  id: string;
  title: string;
  fileName: string;
  objectPath: string;
  fileSize: number;
  amount: number;
  semitones: number;
  use8D: boolean;
  cleanAudio: boolean;
  createdAt: string;
}

export interface SharedAudioInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 1
     * @maxLength 255
     */
  fileName: string;
  /** @pattern ^/objects/uploads/[a-z0-9-]+$ */
  objectPath: string;
  /**
     * @minimum 1
     * @maximum 52428800
     */
  fileSize: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  amount: number;
  /**
     * @minimum -12
     * @maximum 12
     */
  semitones: number;
  use8D: boolean;
  cleanAudio: boolean;
}