# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: sharing
    description: Public audio sharing
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /storage/shares/request-url:
    post:
      operationId: requestShareUploadUrl
      tags: [sharing]
      summary: Request a public audio upload URL
      description: Returns a short-lived upload URL for a processed MP3 or WAV file.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShareUploadRequest"
      responses:
        "201":
          description: Upload URL created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareUploadResponse"
        "400":
          description: Invalid audio metadata
        "500":
          description: Unable to create upload URL
  /storage/tracks:
    get:
      operationId: listSharedAudio
      tags: [sharing]
      summary: List publicly shared processed audio
      description: Returns all processed audio tracks that have been published for public listening.
      responses:
        "200":
          description: Shared audio list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SharedAudio"
        "500":
          description: Unable to list shared audio
    post:
      operationId: createSharedAudio
      tags: [sharing]
      summary: Publish processed audio
      description: Registers an uploaded processed MP3 or WAV file in the public audio library.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SharedAudioInput"
      responses:
        "201":
          description: Shared audio created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedAudio"
        "400":
          description: Invalid shared audio metadata
        "404":
          description: Uploaded audio not found
        "409":
          description: Audio has already been published
        "500":
          description: Unable to publish audio
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ShareUploadRequest:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
          minimum: 1
          maximum: 52428800
        contentType:
          type: string
          enum:
            - audio/mpeg
            - audio/mp3
            - audio/wav
      required:
        - name
        - size
        - contentType
    ShareUploadResponse:
      type: object
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        expiresInSeconds:
          type: integer
      required:
        - uploadURL
        - objectPath
        - expiresInSeconds
    SharedAudio:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        fileName:
          type: string
        objectPath:
          type: string
        fileSize:
          type: integer
        amount:
          type: integer
        semitones:
          type: integer
        use8D:
          type: boolean
        cleanAudio:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - fileName
        - objectPath
        - fileSize
        - amount
        - semitones
        - use8D
        - cleanAudio
        - createdAt
    SharedAudioInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
        fileName:
          type: string
          minLength: 1
          maxLength: 255
        objectPath:
          type: string
          pattern: "^/objects/uploads/[a-z0-9-]+$"
        fileSize:
          type: integer
          minimum: 1
          maximum: 52428800
        amount:
          type: integer
          minimum: 0
          maximum: 100
        semitones:
          type: integer
          minimum: -12
          maximum: 12
        use8D:
          type: boolean
        cleanAudio:
          type: boolean
      required:
        - title
        - fileName
        - objectPath
        - fileSize
        - amount
        - semitones
        - use8D
        - cleanAudio