# lib/api-zod/src/generated/types/shareUploadRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShareUploadRequestContentType } from './shareUploadRequestContentType';

export interface ShareUploadRequest {
  name: string;
  /**
     * @minimum 1
     * @maximum 52428800
     */
  size: number;
  contentType: ShareUploadRequestContentType;
}